Bump chai version
Had some weird state in the container that was causing things to blow up.
Pinpointed the problem to a transient dep of chai. It wasn't the dep's fault per
se, it was more like the dep was updated but chai wasn't. Forcing the image to
be rebuilt with a version bump fixed it. I think a rebuild would've fixed it
without bumping the version but since this is also working I'm comitting it.